Fixed so that current bid is getting updated

+++ backend/src/main/java/com/example/demo/entities/AuctionItem.java
@@ -34,7 +34,7 @@ public class AuctionItem {
     private String images;
     private Boolean sold;
     private Integer startPrice;
-    private Double minimumBid;
+    private Integer minimumBid;
     private Integer primaryImgIndex;
 
     @OneToMany(mappedBy = "auctionItem")
